refactor(editor): GH-1998 extract history mixin and align with Tools
Brings UpstreamDrift's text_editor composition in line with the
canonical Tools structure. Three changes:

1. Extract undo/redo/version-history logic from text_editor.py into
   a new text_editor_history_mixin.py (EditorHistoryMixin). Mirrors
   Tools' text_editor_history_mixin.py module layout.

2. Deduplicate the DbC check blocks across text_editor.py,
   text_editor_diff_mixin.py, and _text_editor_validation.py. A
   prior mechanical `assert X is not None` -> `raise ValueError`
   conversion produced accidentally-duplicated check blocks (e.g.
   two consecutive `if not (X is not None): raise ValueError(...)`
   lines for the same precondition). Collapses the duplicates and
   simplifies the pattern to `if X is None: raise ValueError(...)`.

3. Compose URDFTextEditor from three mixins
   (TextEditorDiffMixin, _URDFValidationMixin, EditorHistoryMixin)
   instead of two. The coordinator class now contains only load/save/
   content-editing/callback responsibilities.

Net effect:
- text_editor.py: 647 -> 445 LOC (-202, ~31%)
- All 149 model_generation tests pass
- ruff check / ruff format / mypy all clean
- No new asserts (issue #2670 policy respected)

Scope: only the text_editor coordinator and its mixins, per
GH-1998's narrow scope. The remaining 300 shared/python/*
filename duplicates are tracked in a separate follow-up issue.

Refs: GH-1998
