+++ src/shared/python/model_generation/editor/text_editor_diff_mixin.py
@@ -61,9 +61,7 @@ class TextEditorDiffMixin:
 
     def _compute_diff(self, original: str, modified: str) -> DiffResult:
         """Compute diff between two strings."""
-        if not (original is not None):
-            raise ValueError("original must be provided")
-        if not (original is not None):
+        if original is None:
             raise ValueError("original must be provided")
         from .text_editor import DiffHunk, DiffResult
 
@@ -158,9 +156,7 @@ class TextEditorDiffMixin:
         Returns:
             List of (left_line, right_line, change_type) tuples.
         """
-        if not (context_lines is not None):
-            raise ValueError("context_lines must be provided")
-        if not (context_lines is not None):
+        if context_lines is None:
             raise ValueError("context_lines must be provided")
         if original is None:
             original = self._original_content  # type: ignore[attr-defined]
